primerykey type changed to Long

+++ src/main/java/com/eshop/repository/ProductRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Collection;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
     // find all product by brand name
     Collection<Product> findAllByBrandName(String brandName);
